+++ kartothek/io/dask/bag_cube.py
@@ -1,9 +1,14 @@
 """
 Dask.Bag IO.
 """
+from typing import Any, Dict, Iterable, Optional
+
 import dask.bag as db
+from simplekv import KeyValueStore
 
 from kartothek.api.discover import discover_datasets_unchecked
+from kartothek.core.cube.cube import Cube
+from kartothek.core.typing import StoreFactory
 from kartothek.io.dask.common_cube import (
     append_to_cube_from_bag_internal,
     build_cube_from_bag_internal,
@@ -22,6 +27,7 @@ from kartothek.io_components.cube.stats import (
     get_metapartitions_for_stats,
     reduce_stats,
 )
+from kartothek.serialization._parquet import ParquetSerializer
 from kartothek.utils.ktk_adapters import get_dataset_keys
 from kartothek.utils.store import copy_keys
 
@@ -39,35 +45,38 @@ __all__ = (
 
 
 def build_cube_from_bag(
-    data,
-    cube,
-    store,
-    ktk_cube_dataset_ids=None,
-    metadata=None,
-    overwrite=False,
-    partition_on=None,
-):
+    data: db.Bag,
+    cube: Cube,
+    store: StoreFactory,
+    ktk_cube_dataset_ids: Optional[Iterable[str]] = None,
+    metadata: Optional[Dict[str, Dict[str, Any]]] = None,
+    overwrite: bool = False,
+    partition_on: Optional[Dict[str, Iterable[str]]] = None,
+    df_serializer: Optional[ParquetSerializer] = None,
+) -> db.Bag:
     """
     Create dask computation graph that builds a cube with the data supplied from a dask bag.
 
     Parameters
     ----------
-    data: dask.Bag
+    data:
         Bag containing dataframes
-    cube: kartothek.core.cube.cube.Cube
+    cube:
         Cube specification.
-    store: Callable[[], simplekv.KeyValueStore]
+    store:
         Store to which the data should be written to.
-    ktk_cube_dataset_ids: Optional[Iterable[str]]
+    ktk_cube_dataset_ids:
         Datasets that will be written, must be specified in advance. If left unprovided, it is assumed that only the
         seed dataset will be written.
-    metadata: Optional[Dict[str, Dict[str, Any]]]
+    metadata:
         Metadata for every dataset.
-    overwrite: bool
+    overwrite:
         If possibly existing datasets should be overwritten.
-    partition_on: Optional[Dict[str, Iterable[str]]]
+    partition_on:
         Optional parition-on attributes for datasets (dictionary mapping :term:`Dataset ID` -> columns).
         See :ref:`Dimensionality and Partitioning Details` for details.
+    df_serializer:
+        Optional Dataframe to Parquet serializer
 
     Returns
     -------
@@ -83,18 +92,20 @@ def build_cube_from_bag(
         metadata=metadata,
         overwrite=overwrite,
         partition_on=partition_on,
+        df_serializer=df_serializer,
     )
 
 
 def extend_cube_from_bag(
-    data,
-    cube,
-    store,
-    ktk_cube_dataset_ids,
-    metadata=None,
-    overwrite=False,
-    partition_on=None,
-):
+    data: db.Bag,
+    cube: Cube,
+    store: KeyValueStore,
+    ktk_cube_dataset_ids: Optional[Iterable[str]],
+    metadata: Optional[Dict[str, Dict[str, Any]]] = None,
+    overwrite: bool = False,
+    partition_on: Optional[Dict[str, Iterable[str]]] = None,
+    df_serializer: Optional[ParquetSerializer] = None,
+) -> db.Bag:
     """
     Create dask computation graph that extends a cube by the data supplied from a dask bag.
 
@@ -102,21 +113,23 @@ def extend_cube_from_bag(
 
     Parameters
     ----------
-    data: dask.Bag
+    data:
         Bag containing dataframes (see :meth:`build_cube` for possible format and types).
-    cube: kartothek.core.cube.cube.Cube
+    cube:
         Cube specification.
-    store: simplekv.KeyValueStore
+    store:
         Store to which the data should be written to.
-    ktk_cube_dataset_ids: Optional[Iterable[str]]
+    ktk_cube_dataset_ids:
         Datasets that will be written, must be specified in advance.
-    metadata: Optional[Dict[str, Dict[str, Any]]]
+    metadata:
         Metadata for every dataset.
-    overwrite: bool
+    overwrite:
         If possibly existing datasets should be overwritten.
-    partition_on: Optional[Dict[str, Iterable[str]]]
+    partition_on:
         Optional parition-on attributes for datasets (dictionary mapping :term:`Dataset ID` -> columns).
         See :ref:`Dimensionality and Partitioning Details` for details.
+    df_serializer:
+        Optional Dataframe to Parquet serializer
 
     Returns
     -------
@@ -132,6 +145,7 @@ def extend_cube_from_bag(
         metadata=metadata,
         overwrite=overwrite,
         partition_on=partition_on,
+        df_serializer=df_serializer,
     )
 
 
@@ -203,7 +217,7 @@ def delete_cube_bag(cube, store, blocksize=100, datasets=None):
     ----------
     cube: Cube
         Cube specification.
-    store: Callable[[], simplekv.KeyValueStore]
+    store: StoreFactory
         KV store.
     blocksize: int
         Number of keys to delete at once.
@@ -246,9 +260,9 @@ def copy_cube_bag(
     ----------
     cube: Cube
         Cube specification.
-    src_store: Callable[[], simplekv.KeyValueStore]
+    src_store: StoreFactory
         Source KV store.
-    tgt_store: Callable[[], simplekv.KeyValueStore]
+    tgt_store: StoreFactory
         Target KV store.
     overwrite: bool
         If possibly existing datasets in the target store should be overwritten.
@@ -342,7 +356,7 @@ def cleanup_cube_bag(cube, store, blocksize=100):
     ----------
     cube: Cube
         Cube specification.
-    store: Union[simplekv.KeyValueStore, Callable[[], simplekv.KeyValueStore]]
+    store: Union[simplekv.KeyValueStore, StoreFactory]
         KV store.
     blocksize: int
         Number of keys to delete at once.
@@ -365,7 +379,14 @@ def cleanup_cube_bag(cube, store, blocksize=100):
     )
 
 
-def append_to_cube_from_bag(data, cube, store, ktk_cube_dataset_ids, metadata=None):
+def append_to_cube_from_bag(
+    data: db.Bag,
+    cube: Cube,
+    store: StoreFactory,
+    ktk_cube_dataset_ids: Optional[Iterable[str]],
+    metadata: Optional[Dict[str, Dict[str, Any]]] = None,
+    df_serializer: Optional[ParquetSerializer] = None,
+) -> db.Bag:
     """
     Append data to existing cube.
 
@@ -383,17 +404,19 @@ def append_to_cube_from_bag(data, cube, store, ktk_cube_dataset_ids, metadata=No
 
     Parameters
     ----------
-    data: dask.Bag
+    data:
         Bag containing dataframes
-    cube: kartothek.core.cube.cube.Cube
+    cube:
         Cube specification.
-    store: Callable[[], simplekv.KeyValueStore]
+    store:
         Store to which the data should be written to.
-    ktk_cube_dataset_ids: Optional[Iterable[str]]
+    ktk_cube_dataset_ids:
         Datasets that will be written, must be specified in advance.
-    metadata: Optional[Dict[str, Dict[str, Any]]]
+    metadata:
         Metadata for every dataset, optional. For every dataset, only given keys are updated/replaced. Deletion of
         metadata keys is not possible.
+    df_serializer:
+        Optional Dataframe to Parquet serializer
 
     Returns
     -------
@@ -407,12 +430,19 @@ def append_to_cube_from_bag(data, cube, store, ktk_cube_dataset_ids, metadata=No
         store=store,
         ktk_cube_dataset_ids=ktk_cube_dataset_ids,
         metadata=metadata,
+        df_serializer=df_serializer,
     )
 
 
 def update_cube_from_bag(
-    data, cube, store, remove_conditions, ktk_cube_dataset_ids, metadata=None
-):
+    data: db.Bag,
+    cube: Cube,
+    store: StoreFactory,
+    remove_conditions,
+    ktk_cube_dataset_ids: Optional[Iterable[str]],
+    metadata: Optional[Dict[str, Dict[str, Any]]] = None,
+    df_serializer: Optional[ParquetSerializer] = None,
+) -> db.Bag:
     """
     Remove partitions and append data to existing cube.
 
@@ -422,20 +452,22 @@ def update_cube_from_bag(
 
     Parameters
     ----------
-    data: dask.Bag
+    data:
         Bag containing dataframes
-    cube: kartothek.core.cube.cube.Cube
+    cube:
         Cube specification.
-    store: Callable[[], simplekv.KeyValueStore]
+    store:
         Store to which the data should be written to.
     remove_conditions
         Conditions that select the partitions to remove. Must be a condition that only uses
         partition columns.
-    ktk_cube_dataset_ids: Optional[Iterable[str]]
+    ktk_cube_dataset_ids:
         Datasets that will be written, must be specified in advance.
-    metadata: Optional[Dict[str, Dict[str, Any]]]
+    metadata:
         Metadata for every dataset, optional. For every dataset, only given keys are updated/replaced. Deletion of
         metadata keys is not possible.
+    df_serializer:
+        Optional Dataframe to Parquet serializer
 
     Returns
     -------
@@ -454,6 +486,7 @@ def update_cube_from_bag(
         remove_conditions=remove_conditions,
         ktk_cube_dataset_ids=ktk_cube_dataset_ids,
         metadata=metadata,
+        df_serializer=df_serializer,
     )