+++ CHANGES.rst
@@ -2,12 +2,14 @@
 Changelog
 =========
 
-Version 3.18.1 (2021-02-XY)
+Version 3.19.0 (2021-02-XY)
 ===========================
 
 * Fix an issue where updates on cubes or updates on datatsets using
   dask.dataframe might not update all secondary indices, resulting in a corrupt
   state after the update
+* Expose compression type and row group chunk size in Cube interface via optional
+  parameter of type :class:`~kartothek.serialization.ParquetSerializer`.
 
 Version 3.18.0 (2021-01-25)
 ===========================
@@ -17,7 +19,7 @@ Version 3.18.0 (2021-01-25)
 * Fix a bug in :func:`~kartothek.io_components.read.dispatch_metapartitions_from_factory` where
   `dispatch_by=[]` would be treated like `dispatch_by=None`, not merging all dataset partitions into
   a single partitions.
-  
+
 Version 3.17.3 (2020-12-04)
 ===========================